Added consolajs to improve UI of console.logs. Renamed pages to views in NextJS example
