+++ core-web/package.json
@@ -95,6 +95,7 @@
         "@tiptap/suggestion": "^2.0.0-beta.218",
         "@webcomponents/webcomponentsjs": "^2.6.0",
         "chart.js": "^4.3.0",
+        "consola": "^3.4.2",
         "core-js": "3.36.1",
         "cross-fetch": "^3.1.4",
         "date-fns": "^2.29.3",
