+++ examples/nextjs/src/views/DetailPage.js
