+++ examples/nextjs/src/views/Page.js
